cv_requirements.txt
docstrange>=1.1.5
openai>=1.0.0
flask>=2.0.0
werkzeug>=2.0.0
python-dotenv>=1.0.0

beautifulsoup4>=4.12.0
docling-ibm-models>=1.0.0
easyocr>=1.7.0
huggingface_hub>=0.19.0
lxml>=4.9.0
markdownify>=0.11.0
mcp>=1.0.0
numpy>=1.24.0
openpyxl>=3.1.0
pandas>=2.0.0
pdf2image>=3.1.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
pypandoc>=1.11.0
python-docx>=0.8.11
python-pptx>=0.6.21
requests>=2.31.0
tiktoken>=0.5.0
tokenizers>=0.15.0
tqdm>=4.66.0
transformers>=4.35.0
